Add basic m3u8 playlist file (#7)
This is basic support for PVR IPTV Simple support in Kodi. It is a poor
man's implementation that (as it is) only adds VRT support (both TV and
Radio) to Kodi.

A better alternative is to use IPTV Manager in the future, nevertheless
this may be useful to some already as-is.